OSIS-4526 move groupelementyear_create views to tree/attach file
