+++ program_management/urls.py
@@ -26,8 +26,9 @@
 from django.conf.urls import url
 from django.urls import include, path
 
+import program_management.views.tree.attach
 import program_management.views.tree.move
-from program_management.views import groupelementyear_create, groupelementyear_delete, groupelementyear_update, \
+from program_management.views import groupelementyear_delete, groupelementyear_update, \
     groupelementyear_read, element_utilization, groupelementyear_postpone, \
     excel, search, tree
 from program_management.views.prerequisite import read, update
@@ -38,16 +39,16 @@ urlpatterns = [
     url(r'^copy_element/$', program_management.views.tree.move.copy_to_cache, name='copy_element'),
     url(r'^(?P<root_id>[0-9]+)/(?P<education_group_year_id>[0-9]+)/', include([
         url(r'^content/', include([
-            url(u'^attach/', groupelementyear_create.PasteElementFromCacheToSelectedTreeNode.as_view(),
+            url(u'^attach/', program_management.views.tree.attach.PasteElementFromCacheToSelectedTreeNode.as_view(),
                 name='education_group_attach'),
-            url(r'^check_attach/', groupelementyear_create.AttachCheckView.as_view(),
+            url(r'^check_attach/', program_management.views.tree.attach.AttachCheckView.as_view(),
                 name="check_education_group_attach"),
-            url(u'^create/$', groupelementyear_create.CreateGroupElementYearView.as_view(),
+            url(u'^create/$', program_management.views.tree.attach.CreateGroupElementYearView.as_view(),
                 name='group_element_year_create'),
             url(r'^(?P<group_element_year_id>[0-9]+)/', include([
                 url(r'^delete/$', groupelementyear_delete.DetachGroupElementYearView.as_view(),
                     name='group_element_year_delete'),
-                url(r'^move/$', groupelementyear_create.MoveGroupElementYearView.as_view(),
+                url(r'^move/$', program_management.views.tree.attach.MoveGroupElementYearView.as_view(),
                     name='group_element_year_move'),
                 url(r'^update/$', groupelementyear_update.UpdateGroupElementYearView.as_view(),
                     name="group_element_year_update"),
